backend/pyproject.toml
[project]
name = "saju-backend"
version = "0.1.0"
description = "SajuBon FastAPI 백엔드 — AI 사주 상담 서비스"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0",
    # LLM / LangChain
    "langchain>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-google-genai>=1.0.0",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",
    # ChromaDB (RAG)
    "chromadb>=0.5",
    "google-genai>=1.0",
    # DB
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    # engine dependencies
    "ephem>=4.1.5",
    "korean-lunar-calendar>=0.3.1",
    "pytz>=2024.1",
]

[dependency-groups]
dev = ["pytest>=8.0", "httpx>=0.27"]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
